Add cloud storage

+++ apps/lupine-notes-and-tools/web/src/components/tools-todo-page.tsx
@@ -11,9 +11,9 @@ import {
 import { ToolItem, LocalToolsService } from '../services/local-tools-service';
 
 export const ToolsTodoPage = (props: {
-  item: ToolItem;
+  item: Partial<ToolItem>;
   sliderFrameHook: SliderFrameHookProps;
-  onEdit?: (item: ToolItem) => void;
+  onEdit?: (item: Partial<ToolItem>) => void;
   onDelete?: (id: string) => void;
 }) => {
   const onEdit = () => {
@@ -27,7 +27,7 @@ export const ToolsTodoPage = (props: {
       cancelButtonText: 'Cancel',
       handleClicked: async (index: number, close: () => void) => {
         close();
-        if (index === 0) {
+        if (index === 0 && props.item.id) {
           LocalToolsService.deleteItem(props.item.id);
           if (props.onDelete) props.onDelete(props.item.id);
           props.sliderFrameHook.close!(new MouseEvent('click'));
@@ -98,7 +98,7 @@ export const ToolsTodoPage = (props: {
         checked={isCompleted}
         disabled={false}
         onClick={(v) => {
-          const cloned = { ...props.item };
+          const cloned = { ...props.item } as ToolItem;
           cloned.todo!.isCompleted = v;
           LocalToolsService.updateItem(cloned);
           props.item.todo!.isCompleted = v;
